refactor: move renderer to lib/
