+++ package.json
@@ -3,8 +3,12 @@
   "version": "1.0.0",
   "description": "Less renderer plugin for Hexo",
   "main": "index.js",
+  "directories": {
+    "lib": "./lib"
+  },
   "files": [
-    "index.js"
+    "index.js",
+    "lib/"
   ],
   "scripts": {
     "eslint": "eslint ."
